adjust package struct of grpc interceptor, move constant of grpc to common package.
